Timing issues resolved concerning comparing user choice array and simon array.
Previously, the game over alert would appear before the full array was compared.
I extended time between the player's playing and the compare function.
This was done by extending the timeout delay at the end of the userPlays function.
Still cannot get the clicker counter to work as an index for iterating so
I can check each play of the user as the user plays.